move url properties to ParserService class
